requirements.txt
Flask==1.1.2
Flask-PyMongo==2.3.0
requests==2.25.1
